pix2pix/requirements.txt
sklearn
psutil
fire
pillow-simd
